Added cookie auth, changed auth settings
Authentification can now be configured as the `auth_method` parameter,
one needs to set it to either `:basic` or `:cookie`
(or nil for no auth); Then, one needs to provide the `:username` and
`:password` options. Because cookie authentification has a timeout,
the cookie will be automatically renewed every
`:cookie_session_minutes` minutes, this option has a default of 9
minutes just to be safe, because the CouchDB default is 10 minutes.